feat: outsource PaneView
